seicross: add cocktail mode (flip screen), ay8910: pass outputs high-impedance state with mem_mask

+++ src/devices/sound/ay8910.h
@@ -308,7 +308,7 @@ private:
 	bool m_active;
 	u8 m_register_latch;
 	u8 m_regs[16 * 2];
-	s16 m_last_enable;
+	u8 m_last_enable;
 	tone_t m_tone[NUM_CHANNELS];
 	envelope_t m_envelope[NUM_CHANNELS];
 	u8 m_prescale_noise;
